backend/requirements.txt
fastapi==0.124.0
uvicorn==0.38.0
google-genai==1.52.0
python-dotenv==1.2.1
pymongo==4.6.0
bcrypt==4.1.2
pydantic[email]==2.12.5
